imdb_scraper/requirements.txt
Scrapy==2.13.3
python-dotenv==1.1.1